voice-agent/pyproject.toml
[project]
name = "voice-agent"
version = "0.1.0"
description = "Spanish voice agent with local STT/LLM/TTS pipeline"
requires-python = ">=3.12"
dependencies = [
    "pipecat-ai[whisper,piper,websocket,local-smart-turn-v3]>=0.0.103",
    "huggingface-hub>=0.36.0",
    "matplotlib>=3.10.8",
    "sounddevice>=0.5.3",
    "openai>=1.0",
    "fastapi>=0.115.0",
    "uvicorn>=0.32.0",
    "websockets>=14.0",
    "numpy>=1.26.0",
    "fastmcp>=2.13",
    "nvidia-cublas-cu12>=12.4",
    "nvidia-cudnn-cu12>=9.0",
    "pytest>=8.0",
    "pytest-asyncio>=0.24",
]

[tool.pytest.ini_options]
asyncio_mode = "auto"
